requirements.txt
langchain
langchain-openai
langchain-anthropic
pydantic
python-dateutil
rich
python-dotenv
